Add match syntax
```
let%hw x y =
  match%hw y with
  | 1 -> 10h
  | 7 -> 3h
  | _ -> 2h
;;
```

Patterns (indices) must be integer constants and the match must include a
(single) wildcard patten.

Indices can be positive and/or negative.

Indices do not need to be consquetive - a tree of mux2 functions is
built to implement the structure.

Right hand side expressions will all be resized to be the same size.

The width of select condition is not fixed - if it is too small for the
total range of indices, only the relevant part of the match will be
built.  The wildcard expression will be used for all unlisted cases.

Also, I have changed how resizing is implemented on the other
operators.  This is to work round the following issue

```
let f x = <some complex expression>
let%hw y = f a + b
```

When rewritten the `f a` expression will be evaluated multiple times
in the resizing code.  This change rebinds it once and resuses the
new binding.